Sync to recent changes from original repo.
* Lost some of the extra changes I did but I don't think they were that
important anyways. Now it's mainly the original version but with new
font and better screen init.
* BlocksDS make file remains unchanged. I do not have this build system
setup right now so I can't attempt adding the custom font graphiocs to
it.
* First palette entry for new icon.bmp changed from pink to black. My
build system seems to produce an icon with pink background due to this.
This is no buano so I had to correct this. I'm not sure how asie is
getting the correct result from this. Since the icon doesn't use pink
anywhere anyways this was a minimal change so shouldn't change the look
of the icon any.
* Added a gitignore. Mostly to prevent dldi files I currently have in my
project folder from showing up on the repo.
* Padding script updated to match new one. Looks like this app uses a
8mb pad file instead of 16MB.